moving uncommon filters to custom folder
